Separates MerinoEditorWindow.cs into separate files for individual classes.
