+++ Assets/Merino/Editor/Scripts/MerinoEditorWindow.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Reflection;
-using System.Text;
 using System.IO;
 using System.Linq;
 using UnityEditor;
@@ -11,7 +10,6 @@ using UnityEngine;
 using Yarn;
 using File = System.IO.File;
 using Merino.EditorCoroutines;
-using Random = System.Random;
 // wow that's a lot of usings
 
 namespace Merino
@@ -1601,372 +1599,4 @@ namespace Merino
 		}
 		#endregion
 	}
-
-	// based a bit on YarnSpinner/ExampleDialogueUI
-	public class MerinoDialogueUI
-	{
-		float textSpeed = 0.01f;
-		public bool inputContinue = false;
-		public int inputOption = -1;
-
-		public string currentNode;
-		string displayString;
-		public string displayStringFull;
-		bool showContinuePrompt = false;
-		string[] optionStrings = new string[0];
-		Yarn.OptionChooser optionChooser;
-		bool useConsolasFont = false;
-
-		public Font consolasFont;
-
-		public IEnumerator RunLine(Yarn.Line line, bool autoAdvance)
-		{
-			displayStringFull = line.text;
-			optionStrings = new string[0];
-			// display dialog
-            if (textSpeed > 0.0f) {
-                // Display the line one character at a time
-                var stringBuilder = new StringBuilder ();
-	            inputContinue = false;
-                foreach (char c in line.text) {
-                    float timeWaited = 0f;
-                    stringBuilder.Append (c);
-                    displayString = stringBuilder.ToString ();
-                    while ( timeWaited < textSpeed )
-                    {
-	                    timeWaited += textSpeed;
-                        // early out / skip ahead
-                        if ( inputContinue ) { break; }
-	                    yield return new WaitForSeconds(timeWaited);
-                    }
-                    if ( inputContinue ) { displayString = line.text; break; }
-                }
-            } else {
-              // Display the line immediately if textSpeed == 0
-                displayString = line.text;
-           }
-
-			inputContinue = false;
-
-            // Show the 'press any key' prompt when done, if we have one
-			showContinuePrompt = true;
-
-            // Wait for any user input
-            while (inputContinue == false && autoAdvance == false)
-            {
-	            yield return new WaitForSeconds(0.01f);
-            }
-
-            // Hide the text and prompt
-			showContinuePrompt = false;
-		}
-		
-		public IEnumerator RunCommand (Yarn.Command command, bool autoAdvance)
-		{
-			optionStrings = new string[0];
-			displayString = "(Yarn command: <<" + command.text + ">>)";
-			inputContinue = false;
-			useConsolasFont = true;
-			showContinuePrompt = true;
-			while (inputContinue == false && autoAdvance == false)
-			{
-				yield return new WaitForSeconds(0.01f);
-			}
-
-			showContinuePrompt = false;
-			useConsolasFont = false;
-		}
-		
-		public void RunOptions (Yarn.Options optionsCollection, Yarn.OptionChooser optionChooser)
-		{
-			// Display each option in a button, and make it visible
-			optionStrings = new string[optionsCollection.options.Count];
-			for(int i=0; i<optionStrings.Length; i++)
-			{
-				optionStrings[i] = optionsCollection.options[i];
-			}
-
-			inputOption = -1;
-			this.optionChooser = optionChooser;
-		}
-		
-//		public IEnumerator NodeComplete(string nextNode)
-//		{
-//			yield return new WaitForSeconds(0.1f);
-//		}
-//		
-//		public IEnumerator DialogueComplete()
-//		{
-//			yield return new WaitForSeconds(0.1f);
-//			displayString = "";
-//		}
-		
-		public void OnGUI(Rect rect)
-		{
-			// background button for clicking to continue			
-			var newRect = new Rect(rect);
-			newRect.y += 20;
-			GUILayout.BeginArea( newRect, EditorStyles.helpBox);
-			// main stuff
-			GUI.enabled = optionStrings.Length == 0;
-			if (GUILayout.Button("", GUILayout.ExpandHeight(true), GUILayout.ExpandWidth(true)))
-			{
-				inputContinue = true;
-			}
-			GUILayout.EndArea();
-			GUI.enabled = true;
-			
-			// display Yarn line here
-			GUILayout.BeginArea( newRect );
-			var passageStyle = new GUIStyle(GUI.skin.box);
-			if (useConsolasFont)
-			{
-				passageStyle.font = consolasFont;
-			}
-			passageStyle.fontSize = 18;
-			passageStyle.normal.textColor = EditorStyles.boldLabel.normal.textColor;
-			passageStyle.padding = new RectOffset(8,8,8,8);
-			passageStyle.richText = true;
-			passageStyle.alignment = TextAnchor.UpperLeft;
-			float maxHeight = passageStyle.CalcHeight(new GUIContent(displayString), rect.width);
-			GUILayout.Label(displayString, passageStyle, GUILayout.Height(0), GUILayout.ExpandHeight(true), GUILayout.MaxHeight(maxHeight), GUILayout.ExpandWidth(true));
-			
-			// show continue prompt
-			if (showContinuePrompt)
-			{
-				Rect promptRect = GUILayoutUtility.GetLastRect();
-				var bounce = Mathf.Sin((float)EditorApplication.timeSinceStartup * 6f) > 0;
-				promptRect.x += promptRect.width-24-(bounce ? 0 : 4);
-				promptRect.y += promptRect.height-24;
-				promptRect.width = 20;
-				promptRect.height = 20;
-				passageStyle.border = new RectOffset(0,0,0,0);
-				passageStyle.padding = new RectOffset(0,0,0,0);
-				passageStyle.alignment = TextAnchor.MiddleCenter;
-				passageStyle.wordWrap = false;
-				passageStyle.normal.background = null;
-				GUI.Box(promptRect, "▶", passageStyle);
-			}
-			
-			// show choices
-			var buttonStyle = new GUIStyle(GUI.skin.button);
-			buttonStyle.fontSize = 16;
-			for (int i = 0; i < optionStrings.Length; i++)
-			{
-				if (GUILayout.Button(optionStrings[i], buttonStyle))
-				{
-					inputOption = i;
-					optionChooser(inputOption);
-					optionStrings = new string[0];
-				}
-			}
-			GUILayout.EndArea();
-		}
-	}
-	
-
-	// this is pretty much just YarnSpinner/ExampleVariableStorage with all the MonoBehaviour stuff taken out
-	public class MerinoVariableStorage : VariableStorage
-	{
-		/// Where we actually keeping our variables
-		Dictionary<string, Yarn.Value> variables = new Dictionary<string, Yarn.Value> ();
-	
-		/// A default value to apply when the object wakes up, or
-		/// when ResetToDefaults is called
-		[System.Serializable]
-		public class DefaultVariable
-		{
-			/// Name of the variable
-			public string name;
-			/// Value of the variable
-			public string value;
-			/// Type of the variable
-			public Yarn.Value.Type type;
-		}
-	
-		/// Our list of default variables, for debugging.
-		DefaultVariable[] defaultVariables = new DefaultVariable[0];
-	
-		/// Reset to our default values when the game starts
-		public MerinoVariableStorage ()
-		{
-			ResetToDefaults ();
-		}
-	
-		/// Erase all variables and reset to default values
-		public void ResetToDefaults ()
-		{
-			Clear ();
-	
-			// For each default variable that's been defined, parse the string
-			// that the user typed in in Unity and store the variable
-			foreach (var variable in defaultVariables) {
-				
-				object value;
-	
-				switch (variable.type) {
-				case Yarn.Value.Type.Number:
-					float f = 0.0f;
-					float.TryParse(variable.value, out f);
-					value = f;
-					break;
-	
-				case Yarn.Value.Type.String:
-					value = variable.value;
-					break;
-	
-				case Yarn.Value.Type.Bool:
-					bool b = false;
-					bool.TryParse(variable.value, out b);
-					value = b;
-					break;
-	
-				case Yarn.Value.Type.Variable:
-					// We don't support assigning default variables from other variables
-					// yet
-					Debug.LogErrorFormat("Can't set variable {0} to {1}: You can't " +
-						"set a default variable to be another variable, because it " +
-						"may not have been initialised yet.", variable.name, variable.value);
-					continue;
-	
-				case Yarn.Value.Type.Null:
-					value = null;
-					break;
-	
-				default:
-					throw new System.ArgumentOutOfRangeException ();
-	
-				}
-	
-				var v = new Yarn.Value(value);
-	
-				SetValue ("$" + variable.name, v);
-			}
-		}
-			
-		/// Set a variable's value
-		public void SetNumber (string variableName, float value)
-		{
-			// Copy this value into our list
-			variables[variableName] = new Yarn.Value(value);
-		}
-	
-		/// Get a variable's value
-		public float GetNumber (string variableName)
-		{
-			// If we don't have a variable with this name, return the null value
-			if (variables.ContainsKey(variableName) == false)
-				return -1f;
-		
-			return variables [variableName].AsNumber;
-		}
-	
-		/// Set a variable's value
-		public void SetValue (string variableName, Yarn.Value value)
-		{
-			// Copy this value into our list
-			variables[variableName] = new Yarn.Value(value);
-		}
-	
-		/// Get a variable's value
-		public Yarn.Value GetValue (string variableName)
-		{
-			// If we don't have a variable with this name, return the null value
-			if (variables.ContainsKey(variableName) == false)
-				return Yarn.Value.NULL;
-			
-			return variables [variableName];
-		}
-	
-		/// Erase all variables
-		public void Clear ()
-		{
-			variables.Clear ();
-		}
-			
-	}
-
-	internal class MerinoMultiColumnHeader : MultiColumnHeader
-	{
-		Mode m_Mode;
-		//Texture helpIcon = EditorGUIUtility.IconContent("_Help").image;
-		
-		public enum Mode
-		{
-			LargeHeader,
-			DefaultHeader,
-			MinimumHeaderWithoutSorting
-		}
-
-		public MerinoMultiColumnHeader(MultiColumnHeaderState state)
-			: base(state)
-		{
-			mode = Mode.DefaultHeader;
-		}
-
-		public Mode mode
-		{
-			get
-			{
-				return m_Mode;
-			}
-			set
-			{
-				m_Mode = value;
-				switch (m_Mode)
-				{
-					case Mode.LargeHeader:
-						canSort = true;
-						height = 37f;
-						break;
-					case Mode.DefaultHeader:
-						canSort = true;
-						height = DefaultGUI.minimumHeight;
-						break;
-					case Mode.MinimumHeaderWithoutSorting:
-						canSort = false;
-						height = DefaultGUI.minimumHeight;
-						break;
-				}
-			}
-		}
-		
-		public override void OnGUI(Rect rect, float xScroll)
-		{
-			// add extra "clear sorting" button if sorting is active
-			var clearSortRect = new Rect(rect);
-			clearSortRect.width = 18;
-			clearSortRect.height = clearSortRect.width;
-			rect.x += clearSortRect.width;
-			rect.width -= clearSortRect.width;
-			if (GUI.Button(clearSortRect, new GUIContent("x", "no sort / clear column sorting"), EditorStyles.miniButton))
-			{
-				state.sortedColumnIndex = -1;
-			}
-			
-			// draw rest of the header
-			base.OnGUI(rect, xScroll);
-		}
-
-		protected override void ColumnHeaderGUI (MultiColumnHeaderState.Column column, Rect headerRect, int columnIndex)
-		{
-			// Default column header gui
-			base.ColumnHeaderGUI(column, headerRect, columnIndex);
-
-			// Add additional info for large header
-			if (mode == Mode.LargeHeader)
-			{
-				// Show example overlay stuff on some of the columns
-				if (columnIndex > 2)
-				{
-					headerRect.xMax -= 3f;
-					var oldAlignment = EditorStyles.largeLabel.alignment;
-					EditorStyles.largeLabel.alignment = TextAnchor.UpperRight;
-					GUI.Label(headerRect, 36 + columnIndex + "%", EditorStyles.largeLabel);
-					EditorStyles.largeLabel.alignment = oldAlignment;
-				}
-			}
-		}
-	}
-
 }
