+++ Assets/Merino/Editor/Scripts/MerinoVariableStorage.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3c42d64e2fe848a89d21de7c13fb8527
+timeCreated: 1536373048
